docs: lifecycle-hooks example - extract things (components, directives) into their own files
This change is in line with Angular Style Guide rule 01-01
https://angular.io/guide/styleguide#style-01-01.
