+++ aio/content/examples/lifecycle-hooks/src/app/after-content-parent.component.ts
@@ -1,0 +1,39 @@
+import { Component } from '@angular/core';
+
+import { LoggerService } from './logger.service';
+
+@Component({
+    selector: 'after-content-parent',
+    template: `
+  <div class="parent">
+    <h2>AfterContent</h2>
+
+    <div *ngIf="show">` +
+        // #docregion parent-template
+        `<after-content>
+        <app-child></app-child>
+      </after-content>`
+        // #enddocregion parent-template
+        + `</div>
+
+    <h4>-- AfterContent Logs --</h4>
+    <p><button (click)="reset()">Reset</button></p>
+    <div *ngFor="let msg of logger.logs">{{msg}}</div>
+  </div>
+  `,
+    styles: ['.parent {background: burlywood}'],
+    providers: [LoggerService]
+})
+export class AfterContentParentComponent {
+    show = true;
+
+    constructor(public logger: LoggerService) {
+    }
+
+    reset() {
+        this.logger.clear();
+        // quickly remove and reload AfterContentComponent which recreates it
+        this.show = false;
+        this.logger.tick_then(() => this.show = true);
+    }
+}
