+++ aio/content/examples/lifecycle-hooks/src/app/after-view-parent.component.ts
@@ -1,0 +1,33 @@
+import { Component } from '@angular/core';
+
+import { LoggerService } from './logger.service';
+
+@Component({
+    selector: 'after-view-parent',
+    template: `
+  <div class="parent">
+    <h2>AfterView</h2>
+
+    <after-view  *ngIf="show"></after-view>
+
+    <h4>-- AfterView Logs --</h4>
+    <p><button (click)="reset()">Reset</button></p>
+    <div *ngFor="let msg of logger.logs">{{msg}}</div>
+  </div>
+  `,
+    styles: ['.parent {background: burlywood}'],
+    providers: [LoggerService]
+})
+export class AfterViewParentComponent {
+    show = true;
+
+    constructor(public logger: LoggerService) {
+    }
+
+    reset() {
+        this.logger.clear();
+        // quickly remove and reload AfterViewComponent which recreates it
+        this.show = false;
+        this.logger.tick_then(() => this.show = true);
+    }
+}
