+++ aio/content/examples/lifecycle-hooks/src/app/child-view.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+// #docregion child-view
+@Component({
+    selector: 'app-child-view',
+    template: '<input [(ngModel)]="hero">'
+})
+export class ChildViewComponent {
+    hero = 'Magneta';
+}
+// #enddocregion child-view
