+++ aio/content/examples/lifecycle-hooks/src/app/child.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-child',
+    template: '<input [(ngModel)]="hero">'
+})
+export class ChildComponent {
+    hero = 'Magneta';
+}
