+++ aio/content/examples/lifecycle-hooks/src/app/do-check-parent.component.ts
@@ -1,0 +1,28 @@
+import { Component, ViewChild } from '@angular/core';
+
+import { DoCheckComponent } from './do-check.component';
+import { Hero } from './hero';
+
+@Component({
+    selector: 'do-check-parent',
+    templateUrl: './do-check-parent.component.html',
+    styles: ['.parent {background: Lavender}']
+})
+export class DoCheckParentComponent {
+    hero: Hero;
+    power: string;
+    title = 'DoCheck';
+    @ViewChild(DoCheckComponent) childView: DoCheckComponent;
+
+    constructor() {
+        this.reset();
+    }
+
+    reset() {
+        this.hero = new Hero('Windstorm');
+        this.power = 'sing';
+        if (this.childView) {
+            this.childView.reset();
+        }
+    }
+}
