+++ aio/content/examples/lifecycle-hooks/src/app/do-check.component.ts
@@ -1,10 +1,8 @@
 /* tslint:disable:forin */
 // #docregion
-import { Component, DoCheck, Input, ViewChild } from '@angular/core';
+import { Component, DoCheck, Input } from '@angular/core';
 
-class Hero {
-  constructor(public name: string) {}
-}
+import { Hero } from './hero';
 
 @Component({
   selector: 'do-check',
@@ -72,22 +70,3 @@ export class DoCheckComponent implements DoCheck {
   }
 }
 
-@Component({
-  selector: 'do-check-parent',
-  templateUrl: './do-check-parent.component.html',
-  styles: ['.parent {background: Lavender}']
-})
-export class DoCheckParentComponent {
-  hero: Hero;
-  power: string;
-  title = 'DoCheck';
-  @ViewChild(DoCheckComponent) childView: DoCheckComponent;
-
-  constructor() { this.reset(); }
-
-  reset() {
-    this.hero = new Hero('Windstorm');
-    this.power = 'sing';
-    if (this.childView) { this.childView.reset(); }
-  }
-}
