+++ aio/content/examples/lifecycle-hooks/src/app/hero.ts
@@ -1,0 +1,4 @@
+export class Hero {
+    constructor(public name: string) {
+    }
+}
