+++ aio/content/examples/lifecycle-hooks/src/app/on-changes-parent.component.ts
@@ -1,0 +1,30 @@
+import { Component, ViewChild } from '@angular/core';
+
+import { OnChangesComponent } from './on-changes.component';
+import { Hero } from './hero';
+
+@Component({
+    selector: 'on-changes-parent',
+    templateUrl: './on-changes-parent.component.html',
+    styles: ['.parent {background: Lavender;}']
+})
+export class OnChangesParentComponent {
+    hero: Hero;
+    power: string;
+    title = 'OnChanges';
+    @ViewChild(OnChangesComponent) childView: OnChangesComponent;
+
+    constructor() {
+        this.reset();
+    }
+
+    reset() {
+        // new Hero object every time; triggers onChanges
+        this.hero = new Hero('Windstorm');
+        // setting power only triggers onChanges if this value is different
+        this.power = 'sing';
+        if (this.childView) {
+            this.childView.reset();
+        }
+    }
+}
