+++ aio/content/examples/lifecycle-hooks/src/app/on-changes.component.ts
@@ -1,13 +1,11 @@
 /* tslint:disable:forin */
 // #docregion
 import {
-  Component, Input, OnChanges,
-  SimpleChanges, ViewChild
+  Component, Input,
+  OnChanges, SimpleChanges
 } from '@angular/core';
 
-class Hero {
-  constructor(public name: string) {}
-}
+import { Hero } from './hero';
 
 @Component({
   selector: 'on-changes',
@@ -46,26 +44,3 @@ export class OnChangesComponent implements OnChanges {
   reset() { this.changeLog = []; }
 }
 
-@Component({
-  selector: 'on-changes-parent',
-  templateUrl: './on-changes-parent.component.html',
-  styles: ['.parent {background: Lavender;}']
-})
-export class OnChangesParentComponent {
-  hero: Hero;
-  power: string;
-  title = 'OnChanges';
-  @ViewChild(OnChangesComponent) childView: OnChangesComponent;
-
-  constructor() {
-    this.reset();
-  }
-
-  reset() {
-    // new Hero object every time; triggers onChanges
-    this.hero = new Hero('Windstorm');
-    // setting power only triggers onChanges if this value is different
-    this.power = 'sing';
-    if (this.childView) { this.childView.reset(); }
-  }
-}
