+++ aio/content/examples/lifecycle-hooks/src/app/peek-a-boo.component.ts
@@ -5,31 +5,17 @@ import {
   AfterContentInit,
   AfterViewChecked,
   AfterViewInit,
-  Directive,
+  Component,
   DoCheck,
+  Input,
   OnChanges,
   OnDestroy,
   OnInit,
   SimpleChanges
 } from '@angular/core';
-import { Component, Input } from '@angular/core';
+
 import { LoggerService } from './logger.service';
-
-let nextId = 1;
-
-// #docregion ngOnInit
-@Directive()
-export class PeekABooDirective implements OnInit {
-  constructor(private logger: LoggerService) { }
-
-  // implement OnInit's `ngOnInit` method
-  ngOnInit() { this.logIt(`OnInit`); }
-
-  logIt(msg: string) {
-    this.logger.log(`#${nextId++} ${msg}`);
-  }
-}
-// #enddocregion ngOnInit
+import { PeekABooDirective } from './peek-a-boo.directive';
 
 @Component({
   selector: 'peek-a-boo',
