+++ aio/content/examples/lifecycle-hooks/src/app/peek-a-boo.directive.ts
@@ -1,0 +1,22 @@
+import { Directive, OnInit } from '@angular/core';
+
+import { LoggerService } from './logger.service';
+
+let nextId = 1;
+
+// #docregion ngOnInit
+@Directive()
+export class PeekABooDirective implements OnInit {
+    constructor(private logger: LoggerService) {
+    }
+
+    // implement OnInit's `ngOnInit` method
+    ngOnInit() {
+        this.logIt(`OnInit`);
+    }
+
+    logIt(msg: string) {
+        this.logger.log(`#${nextId++} ${msg}`);
+    }
+}
+// #enddocregion ngOnInit
